Add frequency analysis script to compare models
To generate quantitative measurements of mesh similarity, this script
generates "feature vectors" by calling buildFeatureVector(path_to_mesh)
that encode low-frequency information about the input mesh. These
feature vectors can be compared (typically using the l2-norm) as a
measurement of how "similar" two meshes are.

This is based on http://nbn-resolving.de/urn:nbn:de:bsz:352-231461, but
since we know that models are partially oriented (the y-axis points
upwards for all models), we can use a simpler rotational invariant that
rotates the model so that the most "concentrated" region of the mesh
points to the front (+z direction).

The FEATURE_VEC_DIM parameter can be adjusted to vary the number of
frequencies included in the feature vector, and the NUM parameter varies
the dimensions of the voxel grid.